+++ CanHazFunny/CanHazFunny.Tests/JokeServiceTests.cs
@@ -12,25 +12,10 @@ namespace CanHazFunny.Tests;
 public class JokeServiceTests
 {
     [Fact]
-    public void GetJoke_NullJoke_ThrowException()
+    public void GetJoke_ReturnsJoke_Success() 
     {
-        // Arrange
-        var mockJokeService = new Mock<IJokeService>();
-        mockJokeService.Setup(js => js.GetJoke()).Throws(new ArgumentNullException());
-
-        // Assert
-        Assert.Throws<ArgumentNullException>(() => mockJokeService.Object.GetJoke());
-    }
-
-    [Fact]
-    public void GetJoke_StandardJoke_Success()
-    {
-        // Arrange
-        string outputJoke = "Very funny joke"!;
-        var mockJokeService = new Mock<IJokeService>();
-        mockJokeService.Setup(js => js.GetJoke()).Returns(outputJoke);
-
-        // Assert
-        Assert.Equal(outputJoke, mockJokeService.Object.GetJoke());
+        var jokeService = new JokeService();
+        var joke = jokeService.GetJoke();
+        Assert.NotNull(joke);
     }
 }
